alteração do sistema para perfil

+++ view/times/excluir_time.php
@@ -1,4 +1,13 @@
 <?php
+// Iniciar a sessão
+session_start();
+
+// Verificar se o professor está logado
+if (!isset($_SESSION['id_professor'])) {
+    // Caso não esteja logado, redireciona para a página de login
+    header('Location: ../login.php');
+    exit();
+}
 
 require '../../app/controller/time.php';
 
@@ -11,6 +20,13 @@ if (isset($_GET['id_time'])) {
     // Chama o método buscar_por_id e recupera o time
     $desc_delete = $objUser->buscar_por_id($id);
 
+    // Verifica se o time pertence ao professor logado
+    if ($desc_delete->id_professor != $_SESSION['id_professor']) {
+        // Caso não seja o professor logado, redireciona para a lista de times
+        header('Location: listar_time.php');
+        exit();
+    }
+
     // Agora chama o método excluirTime passando o ID
     $del = $objUser->excluirTime($id);
 
@@ -21,3 +37,4 @@ if (isset($_GET['id_time'])) {
         echo "Erro ao excluir o time.";
     }
 }
+?>
